:recycle: Break out _record
